requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
coloredlogs==15.0.1
fastembed==0.7.3
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.9.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
loguru==0.7.3
mmh3==5.2.0
mpmath==1.3.0
numpy==2.3.4
onnxruntime==1.23.1
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.33.0
py_rust_stemmers==0.1.5
pydantic==2.12.2
pydantic_core==2.41.4
pypdf==6.1.1
python-dotenv==1.1.1
PyYAML==6.0.3
qdrant-client==1.15.1
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
sentence-transformers==5.1.2
google-genai==1.52.0